requirements.txt
Flask
python-dotenv
Werkzeug
flask-cors
PyJWT
cryptography